// frontend/src/api/api.ts
export interface RejectProductRequest {
    'comment': string;
}
// ...
export class ProductControllerApi extends BaseAPI {
    /**
     * Create product with basic info. Requires X-Username header for audit.
     * @summary Create a new product
     * @param {string} xUsername Username of the user performing the action
     * @param {CreateProductRequest} createProductRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public createProduct(xUsername: string, createProductRequest: CreateProductRequest, options?: RawAxiosRequestConfig) {
        return ProductControllerApiFp(this.configuration).createProduct(xUsername, createProductRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getProduct(id: number, options?: RawAxiosRequestConfig) {
        return ProductControllerApiFp(this.configuration).getProduct(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} code 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getProductByCode(code: string, options?: RawAxiosRequestConfig) {
        return ProductControllerApiFp(this.configuration).getProductByCode(code, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getProductDetail(id: number, options?: RawAxiosRequestConfig) {
        return ProductControllerApiFp(this.configuration).getProductDetail(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} [family] 
     * @param {string} [status] 
     * @param {string} [customerType] 
     * @param {number} [page] 
     * @param {number} [size] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public searchProducts(family?: string, status?: string, customerType?: string, page?: number, size?: number, options?: RawAxiosRequestConfig) {
        return ProductControllerApiFp(this.configuration).searchProducts(family, status, customerType, page, size, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update an existing product
     * @param {number} id Product ID
     * @param {string} xUsername Username of the user performing the action
     * @param {UpdateProductRequest} updateProductRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateProduct(id: number, xUsername: string, updateProductRequest: UpdateProductRequest, options?: RawAxiosRequestConfig) {
        return ProductControllerApiFp(this.configuration).updateProduct(id, xUsername, updateProductRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ProductFeatureControllerApi extends BaseAPI {
    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {ProductFeatureRequest} productFeatureRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public addFeature(productId: number, versionId: number, xUsername: string, productFeatureRequest: ProductFeatureRequest, options?: RawAxiosRequestConfig) {
        return ProductFeatureControllerApiFp(this.configuration).addFeature(productId, versionId, xUsername, productFeatureRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {number} featureId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteFeature(productId: number, versionId: number, featureId: number, options?: RawAxiosRequestConfig) {
        return ProductFeatureControllerApiFp(this.configuration).deleteFeature(productId, versionId, featureId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public listFeatures(productId: number, versionId: number, options?: RawAxiosRequestConfig) {
        return ProductFeatureControllerApiFp(this.configuration).listFeatures(productId, versionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {number} featureId 
     * @param {string} xUsername 
     * @param {ProductFeatureRequest} productFeatureRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateFeature(productId: number, versionId: number, featureId: number, xUsername: string, productFeatureRequest: ProductFeatureRequest, options?: RawAxiosRequestConfig) {
        return ProductFeatureControllerApiFp(this.configuration).updateFeature(productId, versionId, featureId, xUsername, productFeatureRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ProductPricingControllerApi extends BaseAPI {
    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {ProductFeeEntryRequest} productFeeEntryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public addFeeEntry(productId: number, versionId: number, xUsername: string, productFeeEntryRequest: ProductFeeEntryRequest, options?: RawAxiosRequestConfig) {
        return ProductPricingControllerApiFp(this.configuration).addFeeEntry(productId, versionId, xUsername, productFeeEntryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {number} feeEntryId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteFeeEntry(productId: number, versionId: number, feeEntryId: number, options?: RawAxiosRequestConfig) {
        return ProductPricingControllerApiFp(this.configuration).deleteFeeEntry(productId, versionId, feeEntryId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public listFeeEntries(productId: number, versionId: number, options?: RawAxiosRequestConfig) {
        return ProductPricingControllerApiFp(this.configuration).listFeeEntries(productId, versionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {number} feeEntryId 
     * @param {string} xUsername 
     * @param {ProductFeeEntryRequest} productFeeEntryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateFeeEntry(productId: number, versionId: number, feeEntryId: number, xUsername: string, productFeeEntryRequest: ProductFeeEntryRequest, options?: RawAxiosRequestConfig) {
        return ProductPricingControllerApiFp(this.configuration).updateFeeEntry(productId, versionId, feeEntryId, xUsername, productFeeEntryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ProductVersionControllerApi extends BaseAPI {
    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public activate(productId: number, versionId: number, xUsername: string, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).activate(productId, versionId, xUsername, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public approve(productId: number, versionId: number, xUsername: string, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).approve(productId, versionId, xUsername, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} v1 
     * @param {number} v2 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public compareVersions(productId: number, v1: number, v2: number, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).compareVersions(productId, v1, v2, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getVersionDetail(productId: number, versionId: number, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).getVersionDetail(productId, versionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public listVersions(productId: number, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).listVersions(productId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {RejectProductRequest} rejectProductRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public reject(productId: number, versionId: number, xUsername: string, rejectProductRequest: RejectProductRequest, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).reject(productId, versionId, xUsername, rejectProductRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public retire(productId: number, versionId: number, xUsername: string, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).retire(productId, versionId, xUsername, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} productId 
     * @param {number} versionId 
     * @param {string} xUsername 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public submitForApproval(productId: number, versionId: number, xUsername: string, options?: RawAxiosRequestConfig) {
        return ProductVersionControllerApiFp(this.configuration).submitForApproval(productId, versionId, xUsername, options).then((request) => request(this.axios, this.basePath));
    }
}